Fix s1c4 to test for the correct output
The difference between the actual plain text and what I got previously is in
the capitalization. This is a difference in the third highest order bit. In practice
the difference probably doesn't matter since the text is still readable with
reversed casing. However, it does make a difference because spaces and other
non-alphabetic characters are not decripted correctly.